requirements.txt
Django==2.2.5
django-js-asset==1.2.2
django-mptt==0.10.0
django-simple-history==2.7.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
